+++ app/Http/Controllers/MemberController.php
@@ -9,6 +9,8 @@ use Illuminate\Foundation\Auth\SendsPasswordResetEmails;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Password;
 use Illuminate\Support\Facades\Redirect;
+use Spatie\Permission\Models\Permission;
+use Spatie\Permission\Models\Role;
 use Yajra\DataTables\DataTables;
 
 class MemberController extends Controller
@@ -20,6 +22,7 @@ class MemberController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
+
     public function index()
     {
         $members = Member::all();
@@ -33,7 +36,10 @@ class MemberController extends Controller
      */
     public function create()
     {
-        return view('admin.members.create');
+        if (auth()->user()->hasRole('admin')) {
+            return view('admin.members.create');
+        }
+        return view('admin.members.index');
     }
 
     /**
@@ -44,13 +50,15 @@ class MemberController extends Controller
      */
     public function store(MemberRequest $request)
     {
-        Member::query()->create([
+        $member = Member::query()->create([
             'first_name' => $request->first_name,
             'last_name' => $request->last_name,
             'email' => $request->email,
             'cnic' => $request->cnic,
             'address' => $request->address
         ]);
+        $role = Role::updateOrCreate(['name' => 'member']);
+        $member->assignRole($role);
         $response = $this->broker()->sendResetLink(
             $this->credentials($request)
         );
@@ -70,7 +78,15 @@ class MemberController extends Controller
     public function show($id)
     {
         $member = Member::findOrFail($id);
-        return view('admin.members.show', compact('member'));
+        if (auth()->user()->hasRole('admin')) {
+            return view('admin.members.show', compact('member'));
+        }
+        if (auth()->user()->hasRole('member')) {
+            if ($member->id == auth()->user()->id)
+                return view('admin.members.show', compact('member'));
+        }
+        return view('admin.members.index');
+
     }
 
     /**
@@ -81,9 +97,13 @@ class MemberController extends Controller
      */
     public function edit($id)
     {
-        $member = Member::findOrFail($id);
-        return view('admin.members.edit', compact('member'));
+        if (auth()->user()->hasRole('admin')) {
+            $member = Member::findOrFail($id);
+            return view('admin.members.edit', compact('member'));
+        }
+        return view('admin.members.index');
     }
+
 
     /**
      * Update the specified resource in storage.
@@ -92,7 +112,8 @@ class MemberController extends Controller
      * @param  int $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public
+    function update(Request $request, $id)
     {
         $request->validate([
             'first_name' => 'required',
@@ -114,14 +135,21 @@ class MemberController extends Controller
      * @param  int $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public
+    function destroy($id)
     {
-        $member = Member::findOrFail($id);
-        $member->delete();
-        return redirect(route('members.index'));
+        if (auth()->user()->hasRole('admin')) {
+            $member = Member::findOrFail($id);
+            if ($member->id == auth()->user()->id) {
+                $member->delete();
+                return redirect(route('members.index'));
+            }
+        }
+        return view('admin.members.index');
     }
 
-    public function detail($committee_id, $member_id)
+    public
+    function detail($committee_id, $member_id)
     {
         $committee = Committee::findOrFail($committee_id);
         $member = $committee->members()->wherePivot('member_id', $member_id)->first();
@@ -130,7 +158,10 @@ class MemberController extends Controller
 
     public function datatable()
     {
-        $members = Member::select('id', 'first_name', 'last_name', 'email', 'cnic', 'address');
+        if (auth()->user()->hasRole('admin'))
+            $members = Member::select('id', 'first_name', 'last_name', 'email', 'cnic', 'address');
+        else
+            $members = Member::query()->where('id', auth()->user()->id)->select('id', 'first_name', 'last_name', 'email', 'cnic', 'address');
         return Datatables::of($members)
             ->editColumn('first_name', function ($member) {
                 return $member->first_name;
